Add ability to generate and copy SSH keys
With this commit, you can only generate
a key named 'generated', unencrypted,
2048 bits RSA, that overwrites any previous key
named that.  This needs to be a bit better.

(You can only copy the pubkey with the new GUI,
not the private key, because why make it easy
to copy a private key that probably shouldn't
be copied.)